refactor: move db low level interaction into engine module
